FEATURE: parse generics from top file
This is largely motivated by ISE, which barfs if you try to set generics
that are not in the top module. Previously, you would need to make sure
that ALL Hog generics are available in the top project, or it would
generate an error.

With this change, when using ISE only the generics which are actually in
the file get set.